feat: Second Class - Introduction to Windows Forms
Here we have a Windows Forms Application that present us the different types of MessageBox and how we could use it in practice. Also, there are some forms to make specific calcs as: two numbers division, salary adjustment by percentage, example of combobox use and how to make the calc behave in different forms depending on the combobox option selected.